requirements.txt
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
aiosqlite
alembic
amqp
annotated-types
anyio
apache-airflow
apache-airflow-providers-celery
apache-airflow-providers-common-compat
apache-airflow-providers-common-io
apache-airflow-providers-common-sql
apache-airflow-providers-fab
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-smtp
apache-airflow-providers-sqlite
apispec
argcomplete
arize-phoenix
arize-phoenix-evals
arize-phoenix-otel
asgiref
async-timeout
attrs
Authlib
azure-cognitiveservices-speech
babel
beautifulsoup4
billiard
blinker
boto3
botocore
cachelib
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
clickclick
colorama
colorlog
ConfigUpdater
connexion
cron-descriptor
croniter
cryptography
dataclasses-json
decorator
Deprecated
dill
dirtyjson
distro
dnspython
email_validator
exceptiongroup
fastapi
filetype
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
flower
frozenlist
fsspec
google-re2
googleapis-common-protos
graphql-core
greenlet
grpc-interceptor
grpcio
gunicorn
h11
httpcore
httpx
httpx-sse
humanize
idna
imageio
imageio-ffmpeg
importlib-metadata
importlib_resources
inflection
itsdangerous
Jinja2
jiter
jmespath
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kombu
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
lazy-object-proxy
limits
linkify-it-py
llama-cloud
llama-index
llama-index-agent-openai
llama-index-callbacks-arize-phoenix
llama-index-cli
llama-index-core
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-parse
lockfile
lxml
Mako
markdown-it-py
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
methodtools
more-itertools
moviepy
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
openai
openinference-instrumentation
openinference-instrumentation-llama-index
openinference-semantic-conventions
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
orjson
packaging
pandas
pathspec
pendulum
pillow
pluggy
prison
proglog
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
pyarrow
pycparser
pydantic
pydantic-settings
pydantic_core
pydub
Pygments
PyJWT
pymongo
PyMuPDF
pypdf
python-daemon
python-dateutil
python-dotenv
python-multipart
python-nvd3
python-pptx
python-slugify
pytz
PyVimeo
PyYAML
redis
referencing
regex
requests
requests-toolbelt
rfc3339-validator
rich
rich-argparse
rpds-py
s3transfer
scikit-learn
scipy
setproctitle
six
sniffio
soupsieve
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlean.py
sqlparse
starlette
strawberry-graphql
striprtf
tabulate
tenacity
termcolor
text-unidecode
threadpoolctl
tiktoken
time-machine
tinydb
tornado
tqdm
tuspy
typing-inspect
typing_extensions
tzdata
uc-micro-py
universal_pathlib
urllib3
uvicorn
vine
wcwidth
websockets
Werkzeug
wirerope
wrapt
WTForms
XlsxWriter
yarl
zipp
